+++ tests/test_picard_iterations.py
@@ -19,7 +19,7 @@
 
 """
 
-import os
+import pathlib
 
 from fenics import *
 import numpy as np
@@ -28,7 +28,7 @@ import cashocs
 
 rng = np.random.RandomState(300696)
 set_log_level(LogLevel.CRITICAL)
-dir_path = os.path.dirname(os.path.realpath(__file__))
+dir_path = str(pathlib.Path(__file__).parent)
 config = cashocs.load_config(dir_path + "/config_picard.ini")
 
 mesh, subdomains, boundaries, dx, ds, dS = cashocs.regular_mesh(10)